+++ LittleWishWebbackup/src/main/java/pe/edu/upc/aaw/littlewishproject/servicesinterfaces/IExperiencia_LaboralService.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface IExperiencia_LaboralService {
     public void insert(Experiencia_Laboral experiencia_laboral);
     public List<Experiencia_Laboral> list();
-    public void delete(int ID_Experiencia_Laboral);
+    public void delete(Long ID_Experiencia_Laboral);
 }
